(PC-27674)[PRO] refactor: simplify creation components
