--- pro/src/components/VenueForm/CollectiveVenueInformations/CollectiveDmsTimeline/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CollectiveDmsTimeline'
