--- pro/src/components/VenueForm/EACInformation/index.ts
@@ -1,1 +1,0 @@
-export { default as EACInformation } from './EACInformation'
